requirements.txt
yfinance
firebase-admin
newsapi-python
